refactor(sentry): split digest into clustering and scoring modules
Move structural keys/labels into issue_clustering.py and impact scoring
into issue_scoring.py. Deduplicate labels via prefix overrides plus generic
package fallbacks instead of a flat dict that overlapped with elif branches.
